add crud vertex, edge

+++ app/app/models/vertex.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Integer, String
+from sqlalchemy import Boolean, Integer, String, Float
 from sqlalchemy.orm import mapped_column, Mapped, relationship
 
 from app.db.base_class import Base
@@ -8,8 +8,8 @@ class Vertex(Base):
     id: Mapped[int] = mapped_column(
         Integer, primary_key=True, autoincrement=True, index=True
     )
-    x: Mapped[int] = mapped_column(Integer, nullable=True)
-    y: Mapped[int] = mapped_column(Integer, nullable=True)
+    x: Mapped[int] = mapped_column(Float, nullable=True)
+    y: Mapped[int] = mapped_column(Float, nullable=True)
     name: Mapped[str] = mapped_column(String(20), nullable=True)
     description: Mapped[str] = mapped_column(String(50), nullable=True)
 
@@ -21,5 +21,3 @@ class Vertex(Base):
         foreign_keys="[Edge.destination_vertex_id]",
         back_populates="destination_vertex",
     )
-
-
